\name{classifier}
\alias{classifier}
\docType{data}
\title{
An object of class "naiveBayes" generated from data.NaiveBayes
}
\description{
An object of class "naiveBayes" generated from data.NaiveBayes
}
\usage{data("classifier")}
\format{
  An object of class "\code{\link{PASclassifier}}" including components:
 }
\details{
    classifier is generated by 
    
    library(BSgenome.Drerio.UCSC.danRer7)
    
    data(data.NaiveBayes)
    
    classifier <- buildClassifier(data.NaiveBayes$Negative, 
                                  data.NaiveBayes$Positive)
}
\examples{
data(classifier)
names(classifier)
}
\keyword{datasets}
