% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.cleanUpdTSeq-package.R
\docType{package}
\name{cleanUpdTSeq}
\alias{cleanUpdTSeq}
\alias{cleanUpdTSeq-package}
\title{This package classifies putative polyadenylation sites.}
\description{
3'ends of transcripts have generally been poorly annotated. With the advent 
of deep sequencing, many methods have been developed to identify 3'ends. 
The majority of these methods use an oligodT primer which can bind to 
internal adenine-rich sequences, and lead to artifactual identification 
of polyadenylation sites. Heuristic filtering methods rely on a certain 
number of As downstream of a putative polyadenylation site to classify 
the site as true or oligodT primed. This package provides a robust method 
to classify putative polyadenylation sites using a Naive Bayes classifier.
}
\author{
Sarah Sheppard, Jianhong Ou, Nathan Lawson, Lihua Julie Zhu
}
