% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.AllClasses.R
\docType{class}
\name{modelInfo-class}
\alias{modelInfo-class}
\alias{modelInfo}
\alias{$,modelInfo-method}
\alias{$<-,modelInfo-method}
\title{Class \code{"modelInfo"}}
\description{
An object of class \code{"modelInfo"} represents the information of sequence
to use in the analysis
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("modelInfo", upstream, downstream, wordSize, alphabet)}.
}

\keyword{classes}
