% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.AllClasses.R
\docType{class}
\name{naiveBayes-class}
\alias{naiveBayes-class}
\alias{naiveBayes}
\alias{$,naiveBayes-method}
\alias{$<-,naiveBayes-method}
\title{Class \code{"naiveBayes"}}
\description{
An object of class \code{"naiveBayes"} represents the conditional
a-posterior probabilities of a categorical class variable given independent
predictor variables using the Bayes rule.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("naiveBayes", apriori, tables, levels, call)}.
}

\keyword{classes}
