% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.data.R
\docType{data}
\name{classifier}
\alias{classifier}
\title{NaiveBayes classifier}
\format{
An object of class "\code{\link{PASclassifier}}" including
components:
}
\usage{
classifier
}
\description{
An object of class "naiveBayes" generated from data.NaiveBayes
}
\examples{

data(classifier)
names(classifier)

}
\keyword{datasets}
