\name{classifier}
\alias{classifier}
\docType{data}
\title{
An object of class "naiveBayes" generated from data.NaiveBayes
}
\description{
An object of class "naiveBayes" generated from data.NaiveBayes
}
\usage{data("classifier")}
\format{
  An object of class "\code{\link{PASclassifier}}" including components:
 }
\details{
    classifier is generated by 
    data(data.NaiveBayes)
    classifier <- buildClassifier(data.NaiveBayes$Ndata.NaiveBayes, 
                                data.NaiveBayes$Pdata.NaiveBayes)
}
\examples{
data(classifier)
names(classifier)
}
\keyword{datasets}
