% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.data.R
\docType{data}
\name{data.NaiveBayes}
\alias{data.NaiveBayes}
\title{Training Data}
\format{
A list with 2 data frame, "Negative" and "Positive". Negative has
9219 observations on the following 4120 variables. And Positive is a data
frame with 22770 observations on the following 4120 variables.  The format
is: 
\describe{\item{list("Negative")}{'data.frame': 9219 obs. of 4120 variables:}
      \item{list("Positive")}{'data.frame': 22770 obs. of 4120 variables:}}

Both of them have same structure.
\describe{\item{list("y")}{a numeric vector} 
          \item{list("n.A.Downstream")}{a numeric vector}
          \item{list("n.C.Downstream")}{a numeric vector}
          \item{list("n.T.Downstream")}{a numeric vector}
          \item{list("n.G.Downstream")}{a numeric vector}
          \item{list("avg.distanceA2PeakEnd")}{a numeric vector}
          \item{list("dimer")}{a numeric vector}
          \item{: such as AA, AC, AG, AT, CA, ... etc.}{a numeric vector}
          \item{list("heximer")}{a factor with levels \code{0} \code{1}}
          \item{: such as AAAAAA, ACGTAC, ... etc.}{a factor with levels 
               \code{0} \code{1}} 
          \item{list("upstream.seq")}{a vector of sequence string} 
          \item{list("downstream.seq")}{a vector of sequence string}}
}
\usage{
data.NaiveBayes
}
\description{
A RData containing negative and positive training data
}
\examples{
library(BSgenome.Drerio.UCSC.danRer7)
data(data.NaiveBayes)
head(str(data.NaiveBayes$Negative))
head(str(data.NaiveBayes$Positive))
}
\keyword{datasets}
