% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.AllClasses.R
\docType{class}
\name{featureVector-class}
\alias{featureVector-class}
\alias{featureVector}
\alias{$,featureVector-method}
\alias{$<-,featureVector-method}
\title{Class \code{"featureVector"}}
\description{
An object of class \code{"featureVector"} represents the output of
\code{\link{buildFeatureVector}}
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("featureVector", data, info)}.
}

\keyword{classes}
