% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.AllClasses.R
\docType{class}
\name{PASclassifier-class}
\alias{PASclassifier-class}
\alias{PASclassifier}
\alias{$,PASclassifier-method}
\alias{$<-,PASclassifier-method}
\title{Class \code{"PASclassifier"}}
\description{
An object of class \code{"PASclassifier"} represents the output of
\code{\link{buildClassifier}}
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("PASclassifier", classifier, info)}.
}

\examples{

data(classifier)
classifier$info$upstream
classifier$info$wordSize
classifier$info$alphabet

}
\keyword{classes}
